Fall back to original `equal` method
You wouldn’t expect that this is necessary, however `assert.equal` and
the BDD version aren’t actually the same! The BDD version uses strict
equality while the “assert” version doesn’t. TIL.

This fixes the bug with testing equality of an Array and Immutable collection
with the "assert" style.