Improve module loading to make it browser compatible
The chaiImmutable function is now a factory that is given the Immutable
library as an argument.
This is necessary to ensure that it works with the browser (by using
the library already loaded in the global context) and Node.js (by
loading the library with require and using the returned module as the
factory argument).