Remove mocha-phantomjs test suite
At the moment, PhantomJS is not ES6-compatible so a further commit that enables ES6 features is simply breaking it. Furthermore, this is not very useful at the moment, because it merely tests the loading of the plugin in a (headless) browser, nothing more.
I do plan to integrate better testing in browsers, and I do need to re-test the correct loading of these tests in a browser. In the meantime, loading `test/index.html` manually in a browser is a way to do that.